move override_settings to utils
